Start debugging

+++ test/err1/run.sh
@@ -2,6 +2,7 @@
 
 cd ../../goerr/
 go build
+if [ $? -ne 0 ];then exit 1 ;fi
 cd ../test/err1/
 ../../goerr/goerr -f main.go -e err.go merge
 rm err.go
